# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Practice App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:


  /services/categories:
    # binds a127 app logic to a route
    x-swagger-router-controller: services
    get:
      description: Returns a list of all active services
      # used as the method name of the controller
      operationId: servicesCategories
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenericSuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GenericErrorResponse"
            
            
  /services/categories/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: services
    get:
      description: Returns a service category by id
      # used as the method name of the controller
      operationId: servicesCategoriesById
      parameters:
        - in: path
          name: id
          description: Numeric id of the service category to get.
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenericSuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GenericErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GenericSuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  GenericErrorResponse:
    required:
      - message
    properties:
      message:
        type: string